added routes for multiple pages
