+++ src/pages/Favourites.js
@@ -1,1 +1,10 @@
 // List of users favourite movies
+
+function Favorites() {
+
+    return (
+        <h3>This is favorites page</h3>
+    )
+}
+
+export default Favorites
