+++ src/pages/Movie.js
@@ -1,1 +1,10 @@
 // Info for movie
+
+function Movie() {
+
+    return (
+        <h3>Page for single movie</h3>
+    )
+}
+
+export default Movie
