+++ src/pages/Reviews.js
@@ -1,1 +1,10 @@
 // List of reviews
+
+function Reviews() {
+
+    return (
+        <h3>Page for all reviews</h3>
+    )
+}
+
+export default Reviews
